src/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
atlassian-python-api==3.41.19
beautifulsoup4==4.13.3
boto3==1.36.20
botocore==1.36.20
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
fastapi==0.115.8
idna==3.10
jmespath==1.0.1
mangum==0.19.0
oauthlib==3.2.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
wrapt==1.17.2
